Audio configuration implementation in progress. Implemented "Audio" palette which is form containing controls for audio configuration. Implemented "point" and "area" audio types brushes rendering.

+++ source/common_windows.cpp
@@ -737,7 +737,7 @@ void FindDialog::OnKeyDown(wxKeyEvent& event)
 			if(item_list->GetItemCount() > 0) 
 			{
 				size_t n = item_list->GetSelection();
-				if(n == wxNOT_FOUND) 
+				if(n == (size_t) wxNOT_FOUND) 
 					n = 0;
 				else if(n != amount && n - amount < n) // latter is needed for unsigned overflow
 					n -= amount;
@@ -754,7 +754,7 @@ void FindDialog::OnKeyDown(wxKeyEvent& event)
 			if(item_list->GetItemCount() > 0) {
 				size_t n = item_list->GetSelection();
 				size_t itemcount = item_list->GetItemCount();
-				if(n == wxNOT_FOUND)
+				if(n == (size_t) wxNOT_FOUND)
 					n = 0;
 				else if(uint(n) < itemcount - amount && itemcount - amount < itemcount)
 					n += amount;
@@ -1143,7 +1143,7 @@ void ReplaceItemDialog::OnKeyDown(wxKeyEvent& event)
 			if(item_list->GetItemCount() > 0) 
 			{
 				size_t n = item_list->GetSelection();
-				if(n == wxNOT_FOUND) 
+				if(n == (size_t) wxNOT_FOUND) 
 					n = 0;
 				else if(n != amount && n - amount < n) // latter is needed for unsigned overflow
 					n -= amount;
@@ -1160,7 +1160,7 @@ void ReplaceItemDialog::OnKeyDown(wxKeyEvent& event)
 			if(item_list->GetItemCount() > 0) {
 				size_t n = item_list->GetSelection();
 				size_t itemcount = item_list->GetItemCount();
-				if(n == wxNOT_FOUND)
+				if(n == (size_t) wxNOT_FOUND)
 					n = 0;
 				else if(uint(n) < itemcount - amount && itemcount - amount < itemcount)
 					n += amount;
@@ -1313,7 +1313,7 @@ void FindDialogListBox::AddBrush(Brush* brush)
 Brush* FindDialogListBox::GetSelectedBrush() 
 {
 	size_t n = GetSelection();
-	if(n == wxNOT_FOUND || no_matches || cleared)
+	if(n == (size_t) wxNOT_FOUND || no_matches || cleared)
 		return nullptr;
 	return brushlist[n];
 }
